Move cli/pkg/help to sdk/cmd/help to eliminate dependencies on cli package
