+++ docs/generators/cli-doc/main.go
